XMLHttpRequest/xmlhttprequest-network-error-sync.htm

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: members during network errors (sync)</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      test(function() {
        var client = new XMLHttpRequest()
        client.open("GET", "resources/infinite-redirects.php", false)
        assert_throws("NETWORK_ERR", function() { client.send(null) }, "send")
        assert_equals(client.status, 0, "status")
        assert_equals(client.statusText, "", "statusText")
        assert_equals(client.getAllResponseHeaders(), "", "getAllResponseHeaders")
        assert_equals(client.getResponseHeader("content-type"), null, "getResponseHeader")
        assert_equals(client.responseText, "", "responseText")
        assert_equals(client.responseXML, null, "responseXML")
        assert_equals(client.readyState, client.DONE, "readyState")
      })
    </script>
  </body>
</html>
